+++ database/upgrade/upgrade4.7.0RC.sql
@@ -21,6 +21,13 @@ ALTER TABLE myems_system_db.tbl_energy_storage_power_stations DROP COLUMN svg;
 ALTER TABLE myems_system_db.tbl_microgrids ADD `svg_id` BIGINT NOT NULL AFTER `serial_number`;
 ALTER TABLE myems_system_db.tbl_microgrids DROP COLUMN svg;
 
+INSERT INTO myems_system_db.tbl_menus (id,name,route, parent_menu_id,is_hidden)
+VALUES
+(3000,'Work Order','/workorder',NULL,1),
+(3001,'Work Order Installation','/workorder/installation',3000,1),
+(3002,'Work Order Repair','/workorder/repair',3000,1),
+(3003,'Work Order Installation','/workorder/inspection',3000,1);
+
 -- UPDATE VERSION NUMBER
 UPDATE `myems_system_db`.`tbl_versions` SET version='4.7.0RC', release_date='2024-07-07' WHERE id=1;