+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 - added svg actions to myems-api, myems-admin
 - added phase_of_lifecycle to microgird
+- added work order (preview) to myems-web
 ### Changed
 - replaced svg with svg_id in microgrid
 - replaced svg with svg_id in energy storage power station